+++ docs/changelog.md
@@ -21,6 +21,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Fixed
 
 - Fixed mermaid diagrams in component documentation.
+- `six-file-upload`: handleFiles is called twice when a file is dropped.
 
 ## 4.2.6 - 2024-10-04